requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
Pillow>=10.0.0
supabase>=2.0.0
